refactor(purchase-rental-comparison): extract shared types and improve immutability
- Extract ProjectedCosts interface to shared types.ts
- Extract AdvantageList component to dedicated file
- Add Readonly<> to all component props
- Use readonly arrays in constants
- Type PHASE_LABELS with BreakEvenPhase instead of string
- Extract inline prop types to named interfaces
- Disable action buttons as coming soon
