+++ src/components/projects/purchase-rental-comparison/time-horizon-selector.tsx
@@ -26,7 +26,7 @@ export function TimeHorizonSelector({
   onSelect,
   isBreakEvenBeyondMax,
   breakEvenYears,
-}: TimeHorizonSelectorProps) {
+}: Readonly<TimeHorizonSelectorProps>) {
   return (
     <motion.div
       initial={{ opacity: 0, y: 20 }}
