+++ src/components/projects/purchase-rental-comparison/types.ts
@@ -1,0 +1,5 @@
+export interface ProjectedCosts {
+  purchase: number
+  rental: number
+  savings: number
+}
